START move helpers to shared
# Conflicts:
#	test/e2e/overlays.rendering.test.ts
