+++ test/bundles/npm.package.test.ts
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { findFiles } from '../helpers/file-helper';
+import { findFiles } from '../shared/file-helper';
 
 describe('NPM package', () => {
   describe('Files in the dist directory', () => {
