+++ test/e2e/helpers/visu/bpmn-page-utils.ts
@@ -29,7 +29,7 @@ import { BpmnQuerySelectorsForTests } from '@test/shared/query-selectors';
 import { delay } from '../test-utils';
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore js file with commonjs export
-import envUtils = require('../../../helpers/environment-utils.js');
+import envUtils = require('../../../shared/environment-utils.js');
 
 const pageCheckLog = debugLogger('bv:test:page-check');
