+++ test/e2e/overlays.rendering.test.ts
@@ -20,7 +20,7 @@ import type { Page } from 'playwright';
 import { ensureIsArray } from '@lib/component/helpers/array-utils';
 import type { OverlayEdgePosition, OverlayPosition, OverlayShapePosition } from '@lib/component/registry';
 import { ZoomType } from '@lib/component/options';
-import { overlayEdgePositionValues, overlayShapePositionValues } from '../helpers/overlays';
+import { overlayEdgePositionValues, overlayShapePositionValues } from '../shared/overlays';
 import type { Point } from './helpers/visu/bpmn-page-utils';
 import { AvailableTestPages, PageTester } from './helpers/visu/bpmn-page-utils';
 import type { ImageSnapshotThresholdConfig } from './helpers/visu/image-snapshot-config';
